Add an online file input code option to compile
This feature allows passing code as an online URL somewhere, from which
compile.js will pull the text, if possible, and assume it's the code you
want to use. This is done to bypass Discord's 2000 characters limit.

Say you want to run code, that has a bunch of needed classes or
functions, and along with your code it goes over 2000 characters. In
order to run this, you could take your code and the dependencies and
paste it into some online pasting tool, like Hastebin, or Pastebin.
And then pass it to the compiler bot and it'll request that URL, and
execute the code it pulls from that request.

The syntax to use this feature is basically Bash's redirection input
character, meaning, you'd have something like:
`;compile javascript < https://hastebin.com/raw/kixigahite`
This, following rules in Bash, should go before the pipe character |.
If it goes after it will just assume it belongs to the piping and not
pull the code from it.