Fix typo on stdin args of compile.js
At first I named those variables `pipe`, but looking further at the code
I saw it was being referenced as `stdin` which I thought it was a much
more suited name, so I ended up changing it at the last moment. Seems
that when changing it I forgot the old pipe references inside the loop
that forms the stdin and the rest of things, so stdin wasn't ever
fetched correctly.